Allow restoration of Nuget packages for MSBuild projects
Basically just add a second MSBuild project that builds the restore
target. This required adding the target name to the task name, though,
as otherwise they'd be seen as the same and Unimake would bail due to
circular dependencies.

Also fix a minor bug where success was technically spewing an exception
and add VSCode stuff to .gitignore